bail when no new candidates are found
The watcher has to watch a base directory, if we see a changed file we
verify that it matches globs before we try to extract candidates.

If the files doesn't match any of the globs, then we can bail early.
This will result in no visual output which is probably what the user
expects.

However, if a valid file results in no candidates then we also bail
early and don't show anything. Maybe that's confusing, but nothing new
happened and the CSS won't change since it's the same.